requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
beautifulsoup4==4.9.0     # via bs4
bs4==0.0.1                # via -r requirements.in
certifi==2020.4.5.1       # via requests, sentry-sdk
chardet==3.0.4            # via requests
gunicorn==20.0.4          # via -r requirements.in
h-pyramid-sentry==1.2.0   # via -r requirements.in
hupper==1.10.2            # via pyramid
idna==2.9                 # via requests
jinja2==2.11.2            # via pyramid-jinja2
lxml==4.5.0               # via -r requirements.in
markupsafe==1.1.1         # via jinja2, pyramid-jinja2
newrelic==5.12.1.141      # via -r requirements.in
pastedeploy==2.1.0        # via plaster-pastedeploy
plaster-pastedeploy==0.7  # via pyramid
plaster==1.0              # via plaster-pastedeploy, pyramid
pyramid-jinja2==2.8       # via -r requirements.in
pyramid==1.10.4           # via -r requirements.in, h-pyramid-sentry, pyramid-jinja2
pyyaml==5.3.1             # via -r requirements.in
requests==2.23.0          # via -r requirements.in
sentry-sdk==0.14.4        # via h-pyramid-sentry
soupsieve==2.0            # via beautifulsoup4
translationstring==1.3    # via pyramid
urllib3==1.25.9           # via requests, sentry-sdk
venusian==3.0.0           # via pyramid
webob==1.8.6              # via pyramid
whitenoise==5.0.1         # via -r requirements.in
zope.deprecation==4.4.0   # via pyramid, pyramid-jinja2
zope.interface==4.7.2     # via pyramid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
